Refactor file writing out of the discovery module
Until now the `aiven-service-discovery/pkg/discovery` Go module has
done DNS resolution to list Prometheus endpoint IPs in Aiven *and*
written those results to a file.

This commit divides those responsibilities, allowing for other uses to
be made of the endpoint IPs. A new Writer interface and package is
added, and used to preserve the existing behaviour of the
`aiven-service-discovery` BOSH Release.

This change should allow reusing `aiven-service-discovery` when
building other software that interacts with Aiven.